enables from exojax.opacity import Opa(CIA|Rayleigh|Hminus|Mie)
