+++ documents/tutorials/Ackerman_and_Marley_cloud_model.rst
@@ -372,7 +372,7 @@ The Mie scattering can be computed using ``OpaMie``.
     nugrid, wav, res = wavenumber_grid(nus_start, nus_end, N, xsmode="lpf", unit="cm-1")
     
     
-    from exojax.opacity.opacont import OpaMie
+    from exojax.opacity import OpaMie
     
     opa_enstatite = OpaMie(pdb_enstatite, nugrid)
