+++ documents/tutorials/get_started_opart.rst
@@ -54,7 +54,7 @@ Note that you can also use the
     from exojax.database.api  import MdbExomol
     from exojax.database.contdb  import CdbCIA
     from exojax.opacity import OpaPremodit
-    from exojax.opacity.opacont import OpaCIA
+    from exojax.opacity import OpaCIA
     from exojax.rt.layeropacity import single_layer_optical_depth
     from exojax.rt.layeropacity import single_layer_optical_depth_CIA
     from exojax.utils.grids import wavenumber_grid
