+++ documents/tutorials/get_started_transmission.rst
@@ -345,7 +345,7 @@ absorption <https://en.wikipedia.org/wiki/Collision-induced_absorption_and_emiss
 .. code:: ipython3
 
     from exojax.database.contdb  import CdbCIA
-    from exojax.opacity.opacont import OpaCIA
+    from exojax.opacity import OpaCIA
     
     cdb = CdbCIA(".database/H2-H2_2011.cia", nurange=nu_grid)
     opacia = OpaCIA(cdb, nu_grid=nu_grid)
