+++ documents/userguide/customapi.rst
@@ -239,7 +239,7 @@ We will now compare the spectra from ExoMol and Harg2010.
     from exojax.rt.atmrt import ArtEmisPure
     
     from exojax.database.contdb  import CdbCIA
-    from exojax.opacity.opacont import OpaCIA
+    from exojax.opacity import OpaCIA
     
     cdb = CdbCIA(database_dir + "H2-H2_2011.cia", nurange=nu_grid)
