+++ src/exojax/test/generate_methane_spectrum.py
@@ -10,7 +10,7 @@ from exojax.rt.atmrt import ArtEmisPure
 from exojax.database.api  import MdbExomol
 from exojax.opacity import OpaPremodit
 from exojax.database.contdb  import CdbCIA
-from exojax.opacity.opacont import OpaCIA
+from exojax.opacity import OpaCIA
 from exojax.postproc.response import ipgauss_sampling
 from exojax.postproc.spin_rotation import convolve_rigid_rotation
 from exojax.utils.grids import velocity_grid
