+++ tests/endtoend/jaxopt/optimize_spectrum_JAXopt_test.py
@@ -8,7 +8,7 @@ from exojax.database import molinfo
 from exojax.database import contdb 
 from exojax.database.api  import MdbExomol
 from exojax.opacity import OpaDirect
-from exojax.opacity.opacont import OpaCIA
+from exojax.opacity import OpaCIA
 from exojax.rt.atmrt import ArtEmisPure
 from exojax.postproc.specop import SopRotation
 from exojax.postproc.specop import SopInstProfile
