+++ tests/endtoend/reverse/reverse_lpf.py
@@ -12,7 +12,7 @@ from exojax.database import molinfo
 from exojax.utils.grids import wavenumber_grid
 from exojax.utils.instfunc import resolution_to_gaussian_std
 from exojax.opacity import OpaDirect
-from exojax.opacity.opacont import OpaCIA
+from exojax.opacity import OpaCIA
 from exojax.rt.atmrt import ArtEmisPure
 from exojax.postproc.specop import SopRotation
 from exojax.postproc.specop import SopInstProfile
